Refactored index.ts into 3 sources, tests written in ts, updated samples
