+++ browser-wrapper.js
@@ -1,3 +1,3 @@
 // make MicrosoftGraph globally accessible
 // MicrosoftGraph.api() can be called directly
-window.MicrosoftGraph = require('./lib/index.js');
+window.MicrosoftGraph = require('./lib/src/index.js');
