+++ lib/graph-js-sdk-web.js
@@ -1,14 +1,18 @@
 (function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a=typeof require=="function"&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);var f=new Error("Cannot find module '"+o+"'");throw f.code="MODULE_NOT_FOUND",f}var l=n[o]={exports:{}};t[o][0].call(l.exports,function(e){var n=t[o][1][e];return s(n?n:e)},l,l.exports,e,t,n,r)}return n[o].exports}var i=typeof require=="function"&&require;for(var o=0;o<r.length;o++)s(r[o]);return s})({1:[function(require,module,exports){
 // make MicrosoftGraph globally accessible
 // MicrosoftGraph.api() can be called directly
-window.MicrosoftGraph = require('./lib/index.js');
-},{"./lib/index.js":2}],2:[function(require,module,exports){
-'use strict';
+window.MicrosoftGraph = require('./lib/src/index.js');
+},{"./lib/src/index.js":4}],2:[function(require,module,exports){
+"use strict";
+exports.oDataQueryNames = ["$select", "$expand", "$orderby", "$filter", "$top", "$skip", "$skipToken", "$count"];
+exports.DEFAULT_VERSION = "v1.0";
+exports.GRAPH_BASE_URL = "https://graph.microsoft.com/";
+
+},{}],3:[function(require,module,exports){
+"use strict";
+const CommonObj_1 = require("./CommonObj");
 const request = require('superagent');
-const packageInfo = require('../package.json');
-const oDataQueryNames = ["$select", "$expand", "$orderby", "$filter", "$top", "$skip", "$skipToken", "$count"];
-const DEFAULT_VERSION = "v1.0";
-const GRAPH_BASE_URL = "https://graph.microsoft.com/";
+const packageInfo = require('../../package.json');
 class GraphRequest {
     constructor(config, path) {
         this.config = config;
@@ -92,7 +96,7 @@ class GraphRequest {
                 let param = queryParam.split("=");
                 let key = param[0];
                 let value = param[1];
-                if (oDataQueryNames.indexOf(key)) {
+                if (CommonObj_1.oDataQueryNames.indexOf(key)) {
                     this.urlComponents.oDataQueryParams[key] = value;
                 }
                 else {
@@ -235,7 +239,7 @@ class GraphRequest {
                 let request = this.configureRequest(requestBuilder, accessToken);
                 request.end((err, res) => {
                     if (callback !== null)
-                        this.handleResponse(err, res, callback);
+                        GraphRequest.handleResponse(err, res, callback);
                 });
             }
             else if (callback !== null) {
@@ -335,7 +339,7 @@ class GraphRequest {
         }
         return this;
     }
-    handleResponse(err, res, callback) {
+    static handleResponse(err, res, callback) {
         if (res && res.ok) {
             callback(null, res.body, res);
         }
@@ -365,28 +369,198 @@ class GraphRequest {
     }
 }
 exports.GraphRequest = GraphRequest;
-module.exports = class MicrosoftGraphClient {
+
+},{"../../package.json":10,"./CommonObj":2,"superagent":6}],4:[function(require,module,exports){
+'use strict';
+const CommonObj_1 = require("./CommonObj");
+const GraphRequest_1 = require("./GraphRequest");
+class Client {
     constructor() {
         this.config = {
             debugLogging: false,
-            defaultVersion: DEFAULT_VERSION,
-            baseUrl: GRAPH_BASE_URL
+            defaultVersion: CommonObj_1.DEFAULT_VERSION,
+            baseUrl: CommonObj_1.GRAPH_BASE_URL
         };
     }
     static init(clientOptions) {
-        var graphClient = new MicrosoftGraphClient();
+        var graphClient = new Client();
         for (let key in clientOptions) {
             graphClient.config[key] = clientOptions[key];
         }
         return graphClient;
     }
     api(path) {
-        return new GraphRequest(this.config, path);
+        return new GraphRequest_1.GraphRequest(this.config, path);
     }
 }
-;
+exports.Client = Client;
 
-},{"../package.json":8,"superagent":3}],3:[function(require,module,exports){
+},{"./CommonObj":2,"./GraphRequest":3}],5:[function(require,module,exports){
+
+/**
+ * Expose `Emitter`.
+ */
+
+if (typeof module !== 'undefined') {
+  module.exports = Emitter;
+}
+
+/**
+ * Initialize a new `Emitter`.
+ *
+ * @api public
+ */
+
+function Emitter(obj) {
+  if (obj) return mixin(obj);
+};
+
+/**
+ * Mixin the emitter properties.
+ *
+ * @param {Object} obj
+ * @return {Object}
+ * @api private
+ */
+
+function mixin(obj) {
+  for (var key in Emitter.prototype) {
+    obj[key] = Emitter.prototype[key];
+  }
+  return obj;
+}
+
+/**
+ * Listen on the given `event` with `fn`.
+ *
+ * @param {String} event
+ * @param {Function} fn
+ * @return {Emitter}
+ * @api public
+ */
+
+Emitter.prototype.on =
+Emitter.prototype.addEventListener = function(event, fn){
+  this._callbacks = this._callbacks || {};
+  (this._callbacks['$' + event] = this._callbacks['$' + event] || [])
+    .push(fn);
+  return this;
+};
+
+/**
+ * Adds an `event` listener that will be invoked a single
+ * time then automatically removed.
+ *
+ * @param {String} event
+ * @param {Function} fn
+ * @return {Emitter}
+ * @api public
+ */
+
+Emitter.prototype.once = function(event, fn){
+  function on() {
+    this.off(event, on);
+    fn.apply(this, arguments);
+  }
+
+  on.fn = fn;
+  this.on(event, on);
+  return this;
+};
+
+/**
+ * Remove the given callback for `event` or all
+ * registered callbacks.
+ *
+ * @param {String} event
+ * @param {Function} fn
+ * @return {Emitter}
+ * @api public
+ */
+
+Emitter.prototype.off =
+Emitter.prototype.removeListener =
+Emitter.prototype.removeAllListeners =
+Emitter.prototype.removeEventListener = function(event, fn){
+  this._callbacks = this._callbacks || {};
+
+  // all
+  if (0 == arguments.length) {
+    this._callbacks = {};
+    return this;
+  }
+
+  // specific event
+  var callbacks = this._callbacks['$' + event];
+  if (!callbacks) return this;
+
+  // remove all handlers
+  if (1 == arguments.length) {
+    delete this._callbacks['$' + event];
+    return this;
+  }
+
+  // remove specific handler
+  var cb;
+  for (var i = 0; i < callbacks.length; i++) {
+    cb = callbacks[i];
+    if (cb === fn || cb.fn === fn) {
+      callbacks.splice(i, 1);
+      break;
+    }
+  }
+  return this;
+};
+
+/**
+ * Emit `event` with the given args.
+ *
+ * @param {String} event
+ * @param {Mixed} ...
+ * @return {Emitter}
+ */
+
+Emitter.prototype.emit = function(event){
+  this._callbacks = this._callbacks || {};
+  var args = [].slice.call(arguments, 1)
+    , callbacks = this._callbacks['$' + event];
+
+  if (callbacks) {
+    callbacks = callbacks.slice(0);
+    for (var i = 0, len = callbacks.length; i < len; ++i) {
+      callbacks[i].apply(this, args);
+    }
+  }
+
+  return this;
+};
+
+/**
+ * Return array of callbacks for `event`.
+ *
+ * @param {String} event
+ * @return {Array}
+ * @api public
+ */
+
+Emitter.prototype.listeners = function(event){
+  this._callbacks = this._callbacks || {};
+  return this._callbacks['$' + event] || [];
+};
+
+/**
+ * Check if this emitter has `event` handlers.
+ *
+ * @param {String} event
+ * @return {Boolean}
+ * @api public
+ */
+
+Emitter.prototype.hasListeners = function(event){
+  return !! this.listeners(event).length;
+};
+
+},{}],6:[function(require,module,exports){
 /**
  * Root reference for iframes.
  */
@@ -1162,24 +1336,24 @@ Request.prototype.end = function(fn){
   };
 
   // progress
-  var handleProgress = function(e){
+  var handleProgress = function(direction, e) {
     if (e.total > 0) {
       e.percent = e.loaded / e.total * 100;
     }
-    e.direction = 'download';
+    e.direction = direction;
     self.emit('progress', e);
-  };
-  if (this.hasListeners('progress')) {
-    xhr.onprogress = handleProgress;
   }
-  try {
-    if (xhr.upload && this.hasListeners('progress')) {
-      xhr.upload.onprogress = handleProgress;
+  if (this.hasListeners('progress')) {
+    try {
+      xhr.onprogress = handleProgress.bind(null, 'download');
+      if (xhr.upload) {
+        xhr.upload.onprogress = handleProgress.bind(null, 'upload');
+      }
+    } catch(e) {
+      // Accessing xhr.upload fails in IE from a web worker, so just pretend it doesn't exist.
+      // Reported here:
+      // https://connect.microsoft.com/IE/feedback/details/837245/xmlhttprequest-upload-throws-invalid-argument-when-used-from-web-worker-context
     }
-  } catch(e) {
-    // Accessing xhr.upload fails in IE from a web worker, so just pretend it doesn't exist.
-    // Reported here:
-    // https://connect.microsoft.com/IE/feedback/details/837245/xmlhttprequest-upload-throws-invalid-argument-when-used-from-web-worker-context
   }
 
   // timeout
@@ -1364,7 +1538,7 @@ request.put = function(url, data, fn){
   return req;
 };
 
-},{"./is-object":4,"./request":6,"./request-base":5,"emitter":7}],4:[function(require,module,exports){
+},{"./is-object":7,"./request":9,"./request-base":8,"emitter":5}],7:[function(require,module,exports){
 /**
  * Check if `obj` is an object.
  *
@@ -1379,7 +1553,7 @@ function isObject(obj) {
 
 module.exports = isObject;
 
-},{}],5:[function(require,module,exports){
+},{}],8:[function(require,module,exports){
 /**
  * Module of mixed-in functions shared between node and client code
  */
@@ -1458,6 +1632,10 @@ exports.then = function then(resolve, reject) {
   }
   return this._fullfilledPromise.then(resolve, reject);
 }
+
+exports.catch = function(cb) {
+  return this.then(undefined, cb);
+};
 
 /**
  * Allow for extension
@@ -1548,21 +1726,42 @@ exports.unset = function(field){
 };
 
 /**
- * Write the field `name` and `val` for "multipart/form-data"
- * request bodies.
+ * Write the field `name` and `val`, or multiple fields with one object
+ * for "multipart/form-data" request bodies.
  *
  * ``` js
  * request.post('/upload')
  *   .field('foo', 'bar')
  *   .end(callback);
+ *
+ * request.post('/upload')
+ *   .field({ foo: 'bar', baz: 'qux' })
+ *   .end(callback);
  * ```
  *
- * @param {String} name
+ * @param {String|Object} name
  * @param {String|Blob|File|Buffer|fs.ReadStream} val
  * @return {Request} for chaining
  * @api public
  */
 exports.field = function(name, val) {
+
+  // name should be either a string or an object.
+  if (null === name ||  undefined === name) {
+    throw new Error('.field(name, val) name can not be empty');
+  }
+
+  if (isObject(name)) {
+    for (var key in name) {
+      this.field(key, name[key]);
+    }
+    return this;
+  }
+
+  // val should be defined now
+  if (null === val || undefined === val) {
+    throw new Error('.field(name, val) val can not be empty');
+  }
   this._getFormData().append(name, val);
   return this;
 };
@@ -1728,7 +1927,7 @@ exports.send = function(data){
   return this;
 };
 
-},{"./is-object":4}],6:[function(require,module,exports){
+},{"./is-object":7}],9:[function(require,module,exports){
 // The node and browser modules expose versions of this with the
 // appropriate constructor function bound as first argument
 /**
@@ -1762,191 +1961,27 @@ function request(RequestConstructor, method, url) {
 
 module.exports = request;
 
-},{}],7:[function(require,module,exports){
-
-/**
- * Expose `Emitter`.
- */
-
-if (typeof module !== 'undefined') {
-  module.exports = Emitter;
-}
-
-/**
- * Initialize a new `Emitter`.
- *
- * @api public
- */
-
-function Emitter(obj) {
-  if (obj) return mixin(obj);
-};
-
-/**
- * Mixin the emitter properties.
- *
- * @param {Object} obj
- * @return {Object}
- * @api private
- */
-
-function mixin(obj) {
-  for (var key in Emitter.prototype) {
-    obj[key] = Emitter.prototype[key];
-  }
-  return obj;
-}
-
-/**
- * Listen on the given `event` with `fn`.
- *
- * @param {String} event
- * @param {Function} fn
- * @return {Emitter}
- * @api public
- */
-
-Emitter.prototype.on =
-Emitter.prototype.addEventListener = function(event, fn){
-  this._callbacks = this._callbacks || {};
-  (this._callbacks['$' + event] = this._callbacks['$' + event] || [])
-    .push(fn);
-  return this;
-};
-
-/**
- * Adds an `event` listener that will be invoked a single
- * time then automatically removed.
- *
- * @param {String} event
- * @param {Function} fn
- * @return {Emitter}
- * @api public
- */
-
-Emitter.prototype.once = function(event, fn){
-  function on() {
-    this.off(event, on);
-    fn.apply(this, arguments);
-  }
-
-  on.fn = fn;
-  this.on(event, on);
-  return this;
-};
-
-/**
- * Remove the given callback for `event` or all
- * registered callbacks.
- *
- * @param {String} event
- * @param {Function} fn
- * @return {Emitter}
- * @api public
- */
-
-Emitter.prototype.off =
-Emitter.prototype.removeListener =
-Emitter.prototype.removeAllListeners =
-Emitter.prototype.removeEventListener = function(event, fn){
-  this._callbacks = this._callbacks || {};
-
-  // all
-  if (0 == arguments.length) {
-    this._callbacks = {};
-    return this;
-  }
-
-  // specific event
-  var callbacks = this._callbacks['$' + event];
-  if (!callbacks) return this;
-
-  // remove all handlers
-  if (1 == arguments.length) {
-    delete this._callbacks['$' + event];
-    return this;
-  }
-
-  // remove specific handler
-  var cb;
-  for (var i = 0; i < callbacks.length; i++) {
-    cb = callbacks[i];
-    if (cb === fn || cb.fn === fn) {
-      callbacks.splice(i, 1);
-      break;
-    }
-  }
-  return this;
-};
-
-/**
- * Emit `event` with the given args.
- *
- * @param {String} event
- * @param {Mixed} ...
- * @return {Emitter}
- */
-
-Emitter.prototype.emit = function(event){
-  this._callbacks = this._callbacks || {};
-  var args = [].slice.call(arguments, 1)
-    , callbacks = this._callbacks['$' + event];
-
-  if (callbacks) {
-    callbacks = callbacks.slice(0);
-    for (var i = 0, len = callbacks.length; i < len; ++i) {
-      callbacks[i].apply(this, args);
-    }
-  }
-
-  return this;
-};
-
-/**
- * Return array of callbacks for `event`.
- *
- * @param {String} event
- * @return {Array}
- * @api public
- */
-
-Emitter.prototype.listeners = function(event){
-  this._callbacks = this._callbacks || {};
-  return this._callbacks['$' + event] || [];
-};
-
-/**
- * Check if this emitter has `event` handlers.
- *
- * @param {String} event
- * @return {Boolean}
- * @api public
- */
-
-Emitter.prototype.hasListeners = function(event){
-  return !! this.listeners(event).length;
-};
-
-},{}],8:[function(require,module,exports){
+},{}],10:[function(require,module,exports){
 module.exports={
   "name": "msgraph-sdk-javascript",
   "version": "0.1.4",
   "description": "Microsoft Graph JavaScript SDK",
-  "main": "lib/index.js",
-  "typings": "lib/index",
+  "main": "lib/src/index.js",
+  "typings": "lib/src/index",
   "files": [
     "lib/"
   ],
+  "types": "./lib/src/index.d.ts",
   "devDependencies": {
     "typescript": "^1.8.10",
     "browserify": "^13.1.0",
     "mocha":"^3.0.2",
-    "typings": "^1.3.3"
+    "typings": "^2.0.0"
   },
   "scripts": {
     "prepublish": "typings install",
     "build": "tsc && node node-browserify.js > lib/graph-js-sdk-web.js",
-    "test": "mocha test"
+    "test": "mocha lib/test"
   },
   "dependencies": {
     "superagent": "^2.2.0"
