+++ lib/src/CommonObj.js
@@ -1,0 +1,5 @@
+"use strict";
+exports.oDataQueryNames = ["$select", "$expand", "$orderby", "$filter", "$top", "$skip", "$skipToken", "$count"];
+exports.DEFAULT_VERSION = "v1.0";
+exports.GRAPH_BASE_URL = "https://graph.microsoft.com/";
+//# sourceMappingURL=CommonObj.js.map
