+++ lib/src/GraphRequest.js
@@ -1,9 +1,7 @@
-'use strict';
+"use strict";
+const CommonObj_1 = require("./CommonObj");
 const request = require('superagent');
-const packageInfo = require('../package.json');
-const oDataQueryNames = ["$select", "$expand", "$orderby", "$filter", "$top", "$skip", "$skipToken", "$count"];
-const DEFAULT_VERSION = "v1.0";
-const GRAPH_BASE_URL = "https://graph.microsoft.com/";
+const packageInfo = require('../../package.json');
 class GraphRequest {
     constructor(config, path) {
         this.config = config;
@@ -87,7 +85,7 @@ class GraphRequest {
                 let param = queryParam.split("=");
                 let key = param[0];
                 let value = param[1];
-                if (oDataQueryNames.indexOf(key)) {
+                if (CommonObj_1.oDataQueryNames.indexOf(key)) {
                     this.urlComponents.oDataQueryParams[key] = value;
                 }
                 else {
@@ -230,7 +228,7 @@ class GraphRequest {
                 let request = this.configureRequest(requestBuilder, accessToken);
                 request.end((err, res) => {
                     if (callback !== null)
-                        this.handleResponse(err, res, callback);
+                        GraphRequest.handleResponse(err, res, callback);
                 });
             }
             else if (callback !== null) {
@@ -330,7 +328,7 @@ class GraphRequest {
         }
         return this;
     }
-    handleResponse(err, res, callback) {
+    static handleResponse(err, res, callback) {
         if (res && res.ok) {
             callback(null, res.body, res);
         }
@@ -360,24 +358,4 @@ class GraphRequest {
     }
 }
 exports.GraphRequest = GraphRequest;
-module.exports = class MicrosoftGraphClient {
-    constructor() {
-        this.config = {
-            debugLogging: false,
-            defaultVersion: DEFAULT_VERSION,
-            baseUrl: GRAPH_BASE_URL
-        };
-    }
-    static init(clientOptions) {
-        var graphClient = new MicrosoftGraphClient();
-        for (let key in clientOptions) {
-            graphClient.config[key] = clientOptions[key];
-        }
-        return graphClient;
-    }
-    api(path) {
-        return new GraphRequest(this.config, path);
-    }
-}
-;
-//# sourceMappingURL=index.js.map
+//# sourceMappingURL=GraphRequest.js.map
