+++ lib/src/index.d.ts
@@ -1,0 +1,8 @@
+/// <reference path="../../typings/index.d.ts" />
+import { Options } from "./CommonObj";
+import { GraphRequest } from "./GraphRequest";
+export declare class Client {
+    config: Options;
+    static init(clientOptions?: Options): Client;
+    api(path: string): GraphRequest;
+}
