+++ lib/test/responseHandling.d.ts
@@ -1,0 +1,7 @@
+/// <reference path="../../typings/index.d.ts" />
+export declare const OK_SUPERAGENT_RES: {
+    ok: boolean;
+    body: {
+        a: number;
+    };
+};
