+++ lib/test/urlGeneration.d.ts
@@ -1,0 +1,1 @@
+/// <reference path="../../typings/index.d.ts" />
