+++ lib/test/urlParsing.d.ts
@@ -1,0 +1,1 @@
+/// <reference path="../../typings/index.d.ts" />
