+++ lib/test/urlParsing.js
@@ -1,9 +1,7 @@
 "use strict";
-var MicrosoftGraph = require("../lib/index.js");
-var assert = require('assert');
-
-var client = MicrosoftGraph.init({});
-
+const assert = require('assert');
+const index_1 = require("../src/index");
+const client = index_1.Client.init();
 let testCases = {
     "/me": "https://graph.microsoft.com/v1.0/me",
     "/me/": "https://graph.microsoft.com/v1.0/me/",
@@ -16,13 +14,13 @@ let testCases = {
     "/me?a=b": "https://graph.microsoft.com/v1.0/me?a=b",
     "/me?$filter=b&c=d": "https://graph.microsoft.com/v1.0/me?$filter=b&c=d",
     "me?$filter=b&c=d": "https://graph.microsoft.com/v1.0/me?$filter=b&c=d",
-}
-
-describe('#parsePath()', function() {
+};
+describe('#parsePath()', function () {
     for (let path in testCases) {
-        it('should correctly parse ' + path, function() {        
+        it('should correctly parse ' + path, function () {
             var request = client.api(path);
             assert.equal(request.buildFullUrl(), testCases[path]);
         });
     }
-})
+});
+//# sourceMappingURL=urlParsing.js.map
