+++ package.json
@@ -2,21 +2,22 @@
   "name": "msgraph-sdk-javascript",
   "version": "0.1.4",
   "description": "Microsoft Graph JavaScript SDK",
-  "main": "lib/index.js",
-  "typings": "lib/index",
+  "main": "lib/src/index.js",
+  "typings": "lib/src/index",
   "files": [
     "lib/"
   ],
+  "types": "./lib/src/index.d.ts",
   "devDependencies": {
     "typescript": "^1.8.10",
     "browserify": "^13.1.0",
     "mocha":"^3.0.2",
-    "typings": "^1.3.3"
+    "typings": "^2.0.0"
   },
   "scripts": {
     "prepublish": "typings install",
     "build": "tsc && node node-browserify.js > lib/graph-js-sdk-web.js",
-    "test": "mocha test"
+    "test": "mocha lib/test"
   },
   "dependencies": {
     "superagent": "^2.2.0"
