+++ typings.json
@@ -4,7 +4,8 @@
     "microsoft-graph": "registry:common/microsoft-graph#0.2.0+20160924230719"
   },
   "globalDependencies": {
-    "node": "registry:dt/node#6.0.0+20161110151007",
+    "mocha": "registry:dt/mocha#2.2.5+20161028141524",
+    "node": "registry:dt/node#6.0.0+20161121110008",
     "superagent": "registry:dt/superagent#2.0.0+20160906143408"
   }
 }
